+++ src/main/java/com/themetalstorm/bibliothekssystem/controller/UserController.java
@@ -23,7 +23,7 @@ public class UserController {
     }
 
     @GetMapping("/users/{id}")
-    public User getUserById(@PathVariable Long id){
+    public User getUserById(@PathVariable int id){
         return userService.getUserById(id);
     }
